scripts/requirements-ocr.txt
mlx-vlm>=0.3
Pillow>=10.0
pdf2image>=1.17
python-dotenv>=1.0
